# optimize "mul by 3b"