f33059b1be8fee408e156c155dd009bfbff32969 (this PR)
Group operations/double/                                                                            
                        time:   [226.02 µs 227.16 µs 228.61 µs]
Group operations/add/   time:   [322.37 µs 322.60 µs 322.85 µs]                                  

67cc0a2a6bf4fd7f986669504202ca29261bf644 (before this PR)
Group operations/double/                                                                            
                        time:   [235.54 µs 235.64 µs 235.75 µs]
Group operations/add/   time:   [323.36 µs 323.60 µs 323.86 µs]  


093c4dc12bee6ca54a0254534ba2d6497f3b3d67 (using old formula)
Group operations/double/                                                                            
                        time:   [198.25 µs 198.31 µs 198.38 µs]
Group operations/add/   time:   [387.00 µs 387.26 µs 387.54 µs]